isisinsight: ISIS-Fish insight R package

This package aims to provide processes for exploring outputs from the ISIS-Fish (Integration of Spatial Information and Simulation for Fisheries) simulation tool. You can find various types of methods, go for example from visualisation or data manipulation, uncertainty analysis, scenarios display or global exploration.

See also

Author

Maintainer: Mathieu Depetris mathieu.depetris@cnrs.fr (ORCID)

Authors:

Stephanie Mahevas stephanie.mahevas@ifremer.fr

Other contributors:

UMR MARBEC MARine Biodiversity Exploitation & Conservation [copyright holder]